Laravel 11.x Shift (#17384)
* Slim `lang` files

* Remove default `app` files

* Shift core files

* Streamline config files

* Set new `ENV` variables

* Default new `bootstrap/app.php`

* Re-register HTTP middleware

* Consolidate service providers

* Re-register service providers

* Re-register routes

* Re-register scheduled commands

* Bump Composer dependencies

* Adopt anonymous migrations

* Convert `$casts` property to method

* Adopt Laravel type hints

* Mark base controller as `abstract`

* Remove `createApplication` method

* Convert removed migrations methods

* Update composer.lock

* Fix LegacyUserProvider missing method

* Command fixes, might be able to remove some custom code... will check later

* Fix schedule

* database default missing

* Restore artisan name and version using ArtisanStarting event

* Fix middleware names

* Hide commands for lnms in a non-invasive way

* Handle debug output too

* Set public path to html

* fixes

* StyleCI

* fixes

* fixes

* Update phpstan/larastan

* Restore session lifetime for remember me (halved it too)

* restore default, maybe the exceptions is needed too, but this is good for now

* Lock Carbon to v2 for now

* phpstan fixes

* SESSION_SAME_SITE_COOKIE to SESSION_SAME_SITE

* phpstan fixes

* fixes

* Update laravel-console-summary.php to hide more stuff

* remove laravels error handler in tests

* composer.lock

* Flare WIP

* try getting dusk running

* TEST DELETE ME

* rehashPasswordIfRequired: dont run if password not supplied

* fix authenticate is called auth now

* minimal effort to make tests continue

* test fixes

* Update phpunit.xml

* Fix migrations

* Not possible to get a null from currentTranslation because sensor_class is state

* Fix hijacking all errors

* Migration was not fully working

* update test data

* Fix sqlite migrations, they don't need timezone set

* Fix login/out event listeners

* Fix imports

* Give the middleware a good shake out

* remove some unneeded alias

* Add fallback for CACHE_DRIVER and SESSION_SAME_SITE_COOKIE
If we want to remove them, we can remove them in the future.

---------

+++ database/migrations/2018_07_03_091314_create_pollers_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('poller_name')->unique();
             $table->dateTime('last_polled');
             $table->unsignedInteger('devices');
-            $table->float('time_taken', 10, 0);
+            $table->double('time_taken');
         });
     }
